chore: refactor to extract common functionality
