+++ web/src/features/annotation-queues/components/AnnotationQueueItemPage.tsx
@@ -1,44 +1,26 @@
 import { Trace } from "@/src/components/trace";
 import { ObservationPreview } from "@/src/components/trace/ObservationPreview";
 import { TracePreview } from "@/src/components/trace/TracePreview";
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger,
-} from "@/src/components/ui/accordion";
-import { Button } from "@/src/components/ui/button";
 import { Card } from "@/src/components/ui/card";
-import {
-  ResizableHandle,
-  ResizablePanel,
-  ResizablePanelGroup,
-} from "@/src/components/ui/resizable";
 import { Skeleton } from "@/src/components/ui/skeleton";
-import useSessionStorage from "@/src/components/useSessionStorage";
-import { CommentList } from "@/src/features/comments/CommentList";
 import { showSuccessToast } from "@/src/features/notifications/showSuccessToast";
 import { useHasProjectAccess } from "@/src/features/rbac/utils/checkProjectAccess";
-import { AnnotateDrawerContent } from "@/src/features/scores/components/AnnotateDrawerContent";
 import { api } from "@/src/utils/api";
 import { type RouterOutput } from "@/src/utils/types";
 import {
   type AnnotationQueueItem,
   AnnotationQueueObjectType,
   AnnotationQueueStatus,
-  isPresent,
   type ValidatedScoreConfig,
 } from "@langfuse/shared";
-import {
-  ArrowLeft,
-  ArrowRight,
-  SearchXIcon,
-  TriangleAlertIcon,
-} from "lucide-react";
-import { useSession } from "next-auth/react";
+import { ArrowLeft, ArrowRight, SearchXIcon } from "lucide-react";
 import { useRouter } from "next/router";
 import { useEffect, useMemo, useState } from "react";
 import { StringParam, useQueryParam } from "use-query-params";
+import { AnnotationDrawerSection } from "./shared/AnnotationDrawerSection";
+import { AnnotationProcessingLayout } from "./shared/AnnotationProcessingLayout";
+import { useAnnotationQueueData } from "./shared/hooks/useAnnotationQueueData";
+import { Button } from "@/src/components/ui/button";
 
 const AnnotateIOView = ({
   item,
@@ -53,19 +35,8 @@ const AnnotateIOView = ({
   view: "showTree" | "hideTree";
 }) => {
   const router = useRouter();
-  const session = useSession();
   const traceId = item.parentTraceId ?? item.objectId;
   const projectId = router.query.projectId as string;
-  const [showSaving, setShowSaving] = useState(false);
-  const [showComments, setShowComments] = useSessionStorage(
-    `annotationQueueShowComments-${projectId}`,
-    false,
-  );
-  const [panelSize, setPanelSize] = useSessionStorage(
-    `annotationQueuePanelSize-${projectId}`,
-    65,
-  );
-
   const [currentObservationId, setCurrentObservationId] = useQueryParam(
     "observation",
     StringParam,
@@ -79,8 +50,6 @@ const AnnotateIOView = ({
     } else setCurrentObservationId(undefined);
   }, [view, item, setCurrentObservationId]);
 
-  const isLockedByOtherUser = item.lockedByUserId !== session.data?.user?.id;
-
   const trace = api.traces.byIdWithObservationsAndScores.useQuery(
     { traceId, projectId },
     {
@@ -90,10 +59,6 @@ const AnnotateIOView = ({
       },
     },
   );
-
-  const emptySelectedConfigIds = useMemo(() => {
-    return configs.map((c) => c.id);
-  }, [configs]);
 
   if (trace.isLoading || !trace.data)
     return <div className="p-3">Loading...</div>;
@@ -107,128 +72,58 @@ const AnnotateIOView = ({
     isValidObservationId = true;
   }
 
-  return (
-    <ResizablePanelGroup
-      direction="horizontal"
-      className="h-full overflow-hidden"
-      onLayout={(sizes) => {
-        setPanelSize(sizes[0]);
-      }}
-    >
-      <ResizablePanel
-        className="col-span-1 h-full !overflow-y-auto rounded-md border"
-        minSize={30}
-        defaultSize={panelSize}
-      >
-        {view === "hideTree" ? (
-          <div className="max-h-full min-h-0 overflow-y-auto pl-4">
-            {item.objectType === AnnotationQueueObjectType.TRACE ? (
-              <TracePreview
-                key={trace.data.id}
-                trace={trace.data}
-                scores={trace.data.scores}
-                observations={trace.data.observations}
-                viewType="focused"
-              />
-            ) : (
-              <ObservationPreview
-                observations={trace.data.observations}
-                scores={trace.data.scores}
-                projectId={item.projectId}
-                currentObservationId={item.objectId}
-                traceId={traceId}
-                viewType="focused"
-              />
-            )}
-          </div>
+  const leftPanel =
+    view === "hideTree" ? (
+      <div className="max-h-full min-h-0 overflow-y-auto pl-4">
+        {item.objectType === AnnotationQueueObjectType.TRACE ? (
+          <TracePreview
+            key={trace.data.id}
+            trace={trace.data}
+            scores={trace.data.scores}
+            observations={trace.data.observations}
+            viewType="focused"
+          />
         ) : (
-          <div className="max-h-full min-h-0 overflow-y-auto">
-            <Trace
-              key={trace.data.id}
-              trace={trace.data}
-              scores={trace.data.scores}
-              projectId={trace.data.projectId}
-              observations={trace.data.observations}
-              viewType="focused"
-              isValidObservationId={isValidObservationId}
-            />
-          </div>
+          <ObservationPreview
+            observations={trace.data.observations}
+            scores={trace.data.scores}
+            projectId={item.projectId}
+            currentObservationId={item.objectId}
+            traceId={traceId}
+            viewType="focused"
+          />
         )}
-      </ResizablePanel>
-      <ResizableHandle withHandle className="ml-4 bg-transparent" />
-      <ResizablePanel
-        className="col-span-1 h-full md:flex md:flex-col md:overflow-hidden"
-        minSize={30}
-      >
-        <Card className="col-span-2 flex h-full flex-col overflow-hidden">
-          <div className="grid h-full w-full grid-cols-1 grid-rows-[minmax(auto,1fr),minmax(min-content,auto)] justify-between">
-            <div className="w-full overflow-auto">
-              <AnnotateDrawerContent
-                key={"annotation-drawer-content" + item.objectId}
-                scoreTarget={{
-                  type: "trace",
-                  traceId: traceId,
-                  observationId: item.parentTraceId ? item.objectId : undefined,
-                }}
-                scores={trace.data?.scores ?? []}
-                configs={configs}
-                emptySelectedConfigIds={emptySelectedConfigIds}
-                setEmptySelectedConfigIds={() => {}}
-                projectId={item.projectId}
-                analyticsData={{
-                  type: "trace",
-                  source: "AnnotationQueue",
-                }}
-                isSelectHidden
-                queueId={item.queueId}
-                showSaving={showSaving}
-                setShowSaving={setShowSaving}
-                environment={trace.data?.environment}
-                actionButtons={
-                  isLockedByOtherUser && isPresent(item.lockedByUser?.name) ? (
-                    <div className="flex items-center justify-center rounded-sm border border-dark-red bg-light-red p-1">
-                      <TriangleAlertIcon className="mr-1 h-4 w-4 text-dark-red" />
-                      <span className="text-xs text-dark-red">
-                        Currently edited by {item.lockedByUser.name}
-                      </span>
-                    </div>
-                  ) : undefined
-                }
-                isDrawerOpen={true}
-              />
-            </div>
-            <div className="relative max-h-64 overflow-auto">
-              <Accordion
-                type="single"
-                collapsible
-                className="mx-4 mt-4"
-                value={showComments ? "item-1" : ""}
-                onValueChange={(value) => setShowComments(value === "item-1")}
-              >
-                <AccordionItem value="item-1" className="border-none">
-                  <div className="sticky top-0 z-10 border-b bg-background">
-                    <AccordionTrigger
-                      onClick={() => setShowComments(!showComments)}
-                    >
-                      Comments
-                    </AccordionTrigger>
-                  </div>
-                  <AccordionContent>
-                    <CommentList
-                      projectId={item.projectId}
-                      objectId={item.objectId}
-                      objectType={item.objectType}
-                      className="rounded-t-none border-t-transparent"
-                      cardView
-                    />
-                  </AccordionContent>
-                </AccordionItem>
-              </Accordion>
-            </div>
-          </div>
-        </Card>
-      </ResizablePanel>
-    </ResizablePanelGroup>
+      </div>
+    ) : (
+      <div className="max-h-full min-h-0 overflow-y-auto">
+        <Trace
+          key={trace.data.id}
+          trace={trace.data}
+          scores={trace.data.scores}
+          projectId={trace.data.projectId}
+          observations={trace.data.observations}
+          viewType="focused"
+          isValidObservationId={isValidObservationId}
+        />
+      </div>
+    );
+
+  const rightPanel = (
+    <AnnotationDrawerSection
+      item={item}
+      scoreTarget={{
+        type: "trace",
+        traceId: traceId,
+        observationId: item.parentTraceId ? item.objectId : undefined,
+      }}
+      scores={trace.data?.scores ?? []}
+      configs={configs}
+      environment={trace.data?.environment}
+    />
+  );
+
+  return (
+    <AnnotationProcessingLayout leftPanel={leftPanel} rightPanel={rightPanel} />
   );
 };
 
@@ -275,17 +170,7 @@ export const AnnotationQueueItemPage: React.FC<{
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const queueData = api.annotationQueues.byId.useQuery(
-    {
-      queueId: annotationQueueId,
-      projectId,
-    },
-    {
-      refetchOnMount: false,
-      refetchOnWindowFocus: false,
-      refetchOnReconnect: false,
-    },
-  );
+  const { configs } = useAnnotationQueueData({ annotationQueueId, projectId });
 
   const unseenPendingItemCount =
     api.annotationQueueItems.unseenPendingItemCountByQueueId.useQuery(
@@ -327,8 +212,6 @@ export const AnnotationQueueItemPage: React.FC<{
   const totalItems = useMemo(() => {
     return seenItemIds.length + (unseenPendingItemCount.data ?? 0);
   }, [unseenPendingItemCount.data, seenItemIds.length]);
-
-  const configs = queueData.data?.scoreConfigs ?? [];
 
   const relevantItem = useMemo(() => {
     if (isSingleItem) return seenItemData.data;
@@ -380,6 +263,29 @@ export const AnnotationQueueItemPage: React.FC<{
 
   const isNextItemAvailable = totalItems > progressIndex + 1;
 
+  const handleNavigateBack = () => {
+    setProgressIndex(progressIndex - 1);
+  };
+
+  const handleNavigateNext = async () => {
+    if (progressIndex >= seenItemIds.length - 1) {
+      const nextItem = await fetchAndLockNextMutation.mutateAsync({
+        queueId: annotationQueueId,
+        projectId,
+        seenItemIds,
+      });
+      setNextItemData(nextItem);
+    }
+    setProgressIndex(Math.max(progressIndex + 1, 0));
+  };
+
+  const handleComplete = async () => {
+    await completeMutation.mutateAsync({
+      itemId: relevantItem!.id,
+      projectId,
+    });
+  };
+
   return (
     <div className="grid h-full grid-rows-[1fr,auto] gap-4 overflow-hidden">
       {relevantItem ? (
@@ -401,9 +307,7 @@ export const AnnotationQueueItemPage: React.FC<{
               {progressIndex + 1} / {totalItems}
             </span>
             <Button
-              onClick={() => {
-                setProgressIndex(progressIndex - 1);
-              }}
+              onClick={handleNavigateBack}
               variant="outline"
               disabled={progressIndex === 0 || !hasAccess}
               size="lg"
@@ -417,18 +321,8 @@ export const AnnotationQueueItemPage: React.FC<{
         <div className="flex w-full min-w-[265px] justify-end gap-2">
           {!isSingleItem && (
             <Button
-              onClick={async () => {
-                if (progressIndex >= seenItemIds.length - 1) {
-                  const nextItem = await fetchAndLockNextMutation.mutateAsync({
-                    queueId: annotationQueueId,
-                    projectId,
-                    seenItemIds,
-                  });
-                  setNextItemData(nextItem);
-                }
-                setProgressIndex(Math.max(progressIndex + 1, 0));
-              }}
-              disabled={!isNextItemAvailable || !hasAccess} // Disable button during loading
+              onClick={handleNavigateNext}
+              disabled={!isNextItemAvailable || !hasAccess}
               size="lg"
               className={`px-4 ${!relevantItem ? "w-full" : ""}`}
               variant="outline"
@@ -442,12 +336,7 @@ export const AnnotationQueueItemPage: React.FC<{
           {!!relevantItem &&
             (relevantItem.status === AnnotationQueueStatus.PENDING ? (
               <Button
-                onClick={async () => {
-                  await completeMutation.mutateAsync({
-                    itemId: relevantItem.id,
-                    projectId,
-                  });
-                }}
+                onClick={handleComplete}
                 size="lg"
                 className="w-full"
                 disabled={completeMutation.isLoading || !hasAccess}
